+++ paq8pxd.cpp
@@ -127,6 +127,8 @@ inline int max(int a, int b) {return a<b?b:a;}
 #include "prt/mft.hpp"
 #include "models/model.hpp"
 #include "models/sparsematch.hpp"
+#include "models/exe.hpp"
+#include "prt/contextmap2.hpp"
 
 /////////////////////// Global context /////////////////////////
 U8 level=DEFAULT_OPTION;  // Compression level 0 to 15
@@ -645,332 +647,6 @@ int ContextMap::mix1(Mixer& m, int cc, int bp, int c1, int y1) {
 int ContextMap::inputs() {
     return 6;
 }
-template<typename T>
-constexpr auto isPowerOf2(T x) -> bool {
-  return ((x & (x - 1)) == 0);
-}
-/*
-Context map for large contexts (32bits).
-Maps to a bit history state, a 3 MRU byte history, and 1 byte RunStats.
-
-Bit and byte histories are stored in a hash table with 64 byte buckets.
-The buckets are indexed by a context ending after 0, 2 or 5 bits of the
-current byte. Thus, each byte modeled results in 3 main memory accesses
-per context, with all other accesses to cache.
-
-On a byte boundary (bit 0), only 3 of the 7 bit history states are used.
-Of the remaining 4 bytes, 3 are then used to store the last bytes seen
-in this context, 7 bits to store the length of consecutive occurrences of
-the previously seen byte, and 1 bit to signal if more than 1 byte as been
-seen in this context. The byte history is then combined with the bit history
-states to provide additional states that are then mapped to predictions.
-*/
-/*
-   0b 1 1 1 1 1 1 1 1 1   bit set activates mixer input
-      | | | | | | | | |___ short run count
-      | | | | | | | |_____ main run count
-      | | | | | | |_______ main input 1
-      | | | | | |_________ main input 2
-      | | | | |___________ main input 3
-      | | | |_____________ main input 4
-      | | |_______________ main input 5
-      | |_________________ map12
-      |___________________ map6
-  */
-enum {       //  0b111111111
-     CM_RUN1=  0b000000001,
-     CM_RUN0=  0b000000010,
-     CM_MAIN1= 0b000000100,
-     CM_MAIN2= 0b000001000,
-     CM_MAIN3= 0b000010000,
-     CM_MAIN4= 0b000100000,
-     CM_MAIN5= 0b001000000,
-     CM_M12=   0b010000000,
-     CM_M6=    0b100000000,
-     CM_E1=   0b1000000000,
-     CM_E2=  0b10000000000,
-     CM_E3= 0b100000000000,
-     CM_E4=0b1000000000000,
-    CM_E5=0b10000000000000,
- CM_RUN2=0b100000000000000, // RUN2 or RUN0
-  CM_MR=0b1000000000000000
-};
-class ContextMap2 {
-  const U32 C; // max number of contexts
-  class Bucket { // hash bucket, 64 bytes
-    U16 Checksums[7]; // byte context checksums
-    U8 MRU; // last 2 accesses (0-6) in low, high nibble
-  public:
-    U8 BitState[7][7]; // byte context, 3-bit context -> bit history state
-                       // BitState[][0] = 1st bit, BitState[][1,2] = 2nd bit, BitState[][3..6] = 3rd bit
-                       // BitState[][0] is also a replacement priority, 0 = empty
-    inline U8* Find(U16 Checksum) { // Find or create hash element matching checksum.
-                                    // If not found, insert or replace lowest priority (skipping 2 most recent).
-      if (Checksums[MRU&15]==Checksum)
-        return &BitState[MRU&15][0];
-      int worst=0xFFFF, idx=0;
-      for (int i=0; i<7; ++i) {
-        if (Checksums[i]==Checksum)
-          return MRU=MRU<<4|i, (U8*)&BitState[i][0];
-        if (BitState[i][0]<worst && (MRU&15)!=i && MRU>>4!=i) {
-          worst = BitState[i][0];
-          idx=i;
-        }
-      }
-      MRU = 0xF0|idx;
-      Checksums[idx] = Checksum;
-      return (U8*)memset(&BitState[idx][0], 0, 7);
-    }
-  };
-  Array<Bucket, 64> Table; // bit histories for bits 0-1, 2-4, 5-7
-                           // For 0-1, also contains run stats in BitState[][3] and byte history in BitState[][4..6]
-  Array<U8*> BitState; // C pointers to current bit history states
-  Array<U8*> BitState0; // First element of 7 element array containing BitState[i]
-  Array<U8*> ByteHistory; // C pointers to run stats plus byte history, 4 bytes, [RunStats,1..3]
-  Array<U32> Contexts; // C whole byte contexts (hashes)
-  Array<bool> HasHistory; // True if context has a full valid byte history (i.e., seen at least 3 times)
-  StateMap **Maps6b, **Maps8b, **Maps12b, **MapsRun;
-  U32 index; // Next context to set by set()
-  U32 bits;
-  U8 lastByte, lastBit, bitPos;
-  int model;  
-  const int inputCount;
-  const int param;
-  inline void Update() {
-    U64 mask = Table.size()-1;
-    for (U32 i=0; i<index; i++) {
-        if (Contexts[i]){  
-      if (BitState[i])
-        *BitState[i] = nex(*BitState[i], lastBit);
-
-      if (bitPos>1 && ByteHistory[i][0]==0)
-        BitState[i] = nullptr;
-      else {
-        U16 chksum = Contexts[i]>>16;
-        switch (bitPos) {
-          case 0: {
-            BitState[i] = BitState0[i] = Table[(Contexts[i]+bits)&mask].Find(chksum);
-            // Update pending bit histories for bits 2-7
-            if (BitState0[i][3]==2) {
-              const int c = BitState0[i][4]+256;
-              U8 *p = Table[(Contexts[i]+(c>>6))&mask].Find(chksum);
-              p[0] = 1+((c>>5)&1);
-              p[1+((c>>5)&1)] = 1+((c>>4)&1);
-              p[3+((c>>4)&3)] = 1+((c>>3)&1);
-              p = Table[(Contexts[i]+(c>>3))&mask].Find(chksum);
-              p[0] = 1+((c>>2)&1);
-              p[1+((c>>2)&1)] = 1+((c>>1)&1);
-              p[3+((c>>1)&3)] = 1+(c&1);
-              BitState0[i][6] = 0;
-            }
-            // Update byte history of previous context
-            ByteHistory[i][3] = ByteHistory[i][2];
-            ByteHistory[i][2] = ByteHistory[i][1];
-            if (ByteHistory[i][0]==0)  // new context
-              ByteHistory[i][0]=2, ByteHistory[i][1]=lastByte;
-            else if (ByteHistory[i][1]!=lastByte)  // different byte in context
-              ByteHistory[i][0]=1, ByteHistory[i][1]=lastByte;
-            else if (ByteHistory[i][0]<254)  // same byte in context
-              ByteHistory[i][0]+=2;
-            else if (ByteHistory[i][0]==255) // more than one byte seen, but long run of current byte, reset to single byte seen
-              ByteHistory[i][0] = 128;
-
-            ByteHistory[i] = BitState0[i]+3;
-            HasHistory[i] = *BitState0[i]>15;
-            break;
-          }
-          case 2: case 5: {
-            BitState[i] = BitState0[i] = Table[(Contexts[i]+bits)&mask].Find(chksum);
-            break;
-          }
-          case 1: case 3: case 6: BitState[i] = BitState0[i]+1+lastBit; break;
-          case 4: case 7: BitState[i] = BitState0[i]+3+(bits&3); break;
-        }
-      }
-      }
-    }
-  }
-public:
-  // Construct using Size bytes of memory for Count contexts
-  ContextMap2(const U64 Size, const U32 Count,int mod,int m=0b110111110) :
-   C(Count), Table(Size>>6), BitState(Count), BitState0(Count), ByteHistory(Count),
-    Contexts(Count), HasHistory(Count),model(mod),inputCount(__builtin_popcount(m)),param(m) {
-    assert(Size>=64 && isPowerOf2(Size));
-    assert(sizeof(Bucket)==64);
-    if (param &CM_M6) Maps6b = new StateMap*[C];
-    Maps8b = new StateMap*[C];
-    if (param &CM_M12) Maps12b = new StateMap*[C];
-    if (param &CM_MR) MapsRun = new StateMap*[C];
-    for (U32 i=0; i<C; i++) {
-      if(param &CM_M6) Maps6b[i] = new StateMap((1<<6)+8);
-      Maps8b[i] = new StateMap(1<<8);
-      if (param&CM_M12) Maps12b[i] = new StateMap((1<<12)+(1<<9));
-      if (param &CM_MR) MapsRun[i] = new StateMap((1<<12));
-      BitState[i] = BitState0[i] = &Table[i].BitState[0][0];
-      ByteHistory[i] = BitState[i]+3;
-    }
-    index = 0;
-    lastByte = lastBit = 0;
-    bits = 1;  bitPos = 0;
-#ifdef VERBOSE
-    printf("Model: %s\n",modelNames[model].c_str());
-    printf("  CM total inputs: %d\n",inputCount);
-    printf("                   MR R2 R1 R0 M1 M2 M3 M4 M5 M12 M6 E1 E2 E3 E4 E5 \n");
-    printf("  CM active inputs  %d  %d  %d  %d  %d  %d  %d  %d  %d   %d  %d  %d  %d  %d  %d  %d\n",
-    param&CM_MR?1:0,param&CM_RUN2?1:0,param&CM_RUN1?1:0, param&CM_RUN0?1:0, 
-    param&CM_MAIN1?1:0, param&CM_MAIN2?1:0, param&CM_MAIN3?1:0, param&CM_MAIN4?1:0, param&CM_MAIN5?1:0,
-    param&CM_M12?1:0, param&CM_M6?1:0,
-    param&CM_E1?1:0, param&CM_E2?1:0, param&CM_E3?1:0, param&CM_E4?1:0, param&CM_E5?1:0
-    );
-#endif
-  }
-  ~ContextMap2() {
-    for (U32 i=0; i<C; i++) {
-      if (param &CM_M6) delete Maps6b[i];
-      delete Maps8b[i];
-      if (param&CM_M12) delete Maps12b[i];
-      if (param &CM_MR) delete MapsRun[i];
-    }
-    if (param &CM_M6) delete[] Maps6b;
-    delete[] Maps8b;
-    if (param&CM_M12) delete[] Maps12b;
-    if (param &CM_MR) delete[] MapsRun;
-  }
-
-  inline void set(U32 ctx) { // set next whole byte context to ctx
-  //assert(index>0 && index<=C); // fail if assert on
-  if (index==C) index=0;     // model bypass, FIXME
-    if (ctx==0){ 
-      Contexts[index]=0; 
-    } else{
-      ctx = ctx*987654323+index; // permute (don't hash) ctx to spread the distribution
-      ctx = ctx<<16|ctx>>16;
-      Contexts[index] = ctx*123456791+index;
-    }
-    index++;
-  }
-  inline void  set() {
-    set(U32(0));
-}
-  int mix(Mixer& m, const int Multiplier = 1, const int Divisor = 4) {
-    int result = 0;
-    lastBit = m.x.y;
-    bitPos = m.x.bpos;
-    bits+=bits+lastBit;
-    lastByte = bits&0xFF;
-    if (bitPos==0)
-      bits = 1;
-
-    Update();
-
-    for (U32 i=0; i<index; i++) {
-      if (Contexts[i]){  
-      // predict from bit context
-      int state = (BitState[i])?*BitState[i]:0;
-      int p1=0 ;
-      if (state){
-          if (m.x.count>0x5FFFF) p1=Maps8b[i]->p1(state,m.x.y);else p1=Maps8b[i]->p(state,m.x.y);
-      } 
-      int n0=nex(state, 2), n1=nex(state, 3), k=-~n1;
-      const int bitIsUncertain = int(n0 != 0 && n1 != 0);
-      k = (k*64)/(k-~n0);
-      n0=-!n0, n1=-!n1;
-
-      // predict from last byte in context
-      if ((U32)((ByteHistory[i][1]+256)>>(8-bitPos))==bits){
-        int RunStats = ByteHistory[i][0]; // count*2, +1 if 2 different bytes seen
-        const int predictedBit = (ByteHistory[i][1] >> (7 - bitPos)) & 1U;
-        int sign=(predictedBit)*2-1;  // predicted bit + for 1, - for 0
-        int value = ilog(RunStats+1)<<(3-(RunStats&1));
-        const int byte1IsUncertain = static_cast<const int>(ByteHistory[i][2] != ByteHistory[i][1]);
-        const int bp = (0xFEA4U >> (bitPos << 1U)) & 3U; // {0}->0  {1}->1  {2,3,4}->2  {5,6,7}->3
-        if (param&CM_RUN1)      m.add(sign * (min(RunStats>>1, 32) << 5)); // +/- 32..1024
-        if (param&CM_RUN0)      m.add(sign*value);
-        if (param&CM_RUN2)      m.add(sign*value);
-        if (param&CM_MR)        m.add(stretch(MapsRun[i]->p( (RunStats>>1) << 4U | bp << 2U | byte1IsUncertain << 1 | predictedBit,m.x.y)) >> (1 + byte1IsUncertain));
-      }
-      else if (bitPos>0 && (ByteHistory[i][0]&1)>0) {
-        if  (param&CM_RUN1)     m.add(0);
-        if  (param&CM_RUN2)     m.add(0);
-        if  (param&CM_MR)       m.add(0);
-        if ((U32)((ByteHistory[i][2]+256)>>(8-bitPos))==bits)
-          {if (param&CM_RUN0)   m.add((((ByteHistory[i][2]>>(7-bitPos))&1)*2-1)*128);}
-        else if (HasHistory[i] && (U32)((ByteHistory[i][3]+256)>>(8-bitPos))==bits)
-          {if (param&CM_RUN0)   m.add((((ByteHistory[i][3]>>(7-bitPos))&1)*2-1)*128);}
-        else
-          {if (param&CM_RUN0)   m.add(0);}
-      }
-      else{
-       if (param&CM_RUN2)       m.add(0);
-       if (param&CM_RUN1)       m.add(0);
-       if (param&CM_RUN0)       m.add(0);
-       if (param&CM_MR)         m.add(0);
-      }
-      if (state == 0) {
-            if (param&CM_MAIN1) m.add(0);
-            if (param&CM_MAIN2) m.add(0);
-            if (param&CM_MAIN3) m.add(0);
-            if (param&CM_MAIN4) m.add(0);
-            if (param&CM_MAIN5) m.add(0);
-            if (param&CM_E1) m.add(0);
-            if (param&CM_E2) m.add(0);
-            if (param&CM_E3) m.add(0);
-            if (param&CM_E4) m.add(0);
-            if (param&CM_E5) m.add(0);
-      } else {
-        const int contextIsYoung = int(state <= 2);
-        int st=(stretch(p1)*Multiplier)/Divisor;
-        if (param&CM_MAIN1) m.add(st >> contextIsYoung);
-        if (param&CM_MAIN2) m.add(((p1-2048)*Multiplier)/(2*Divisor));
-        if (param&CM_MAIN3) m.add((bitIsUncertain - 1) & st); // when both counts are nonzero add(0) otherwise add(st)
-         int p0=4095-p1;
-        if (param&CM_MAIN4) m.add((((p1&n0)-(p0&n1))*Multiplier)/(4*Divisor));
-        if (param&CM_MAIN5) m.add(0);//
-        if (param&CM_E1) m.add(0);// 
-        if (param&CM_E2) m.add(0);
-        if (param&CM_E3) m.add(0);
-        if (param&CM_E4) m.add(0);
-        if (param&CM_E5) m.add(0);
-        result++;
-      }
-
-      if (HasHistory[i]) {
-        state  = (ByteHistory[i][1]>>(7-bitPos))&1;
-        state |= ((ByteHistory[i][2]>>(7-bitPos))&1)*2;
-        state |= ((ByteHistory[i][3]>>(7-bitPos))&1)*4;
-      }
-      else
-        state = 8;
-
-      if (param&CM_M12)  m.add(stretch(Maps12b[i]->p((state<<9)|(bitPos<<6)|k,m.x.y) )>>2);
-      if (param&CM_M6)   m.add(stretch(Maps6b[i]->p((state<<3)|bitPos,m.x.y))>>2);
-    
-      } else{
-        if (param&CM_RUN2)  m.add(0);
-        if (param&CM_RUN1)  m.add(0);
-        if (param&CM_RUN0)  m.add(0);
-        if (param&CM_MAIN1) m.add(0);
-        if (param&CM_MAIN2) m.add(0);
-        if (param&CM_MAIN3) m.add(0);
-        if (param&CM_MAIN4) m.add(0);
-        if (param&CM_MAIN5) m.add(0);//
-        if (param&CM_M12)   m.add(0);
-        if (param&CM_M6)    m.add(0);
-        if (param&CM_E1) m.add(0);//
-        if (param&CM_E2) m.add(0);//
-        if (param&CM_E3) m.add(0);//
-        if (param&CM_E4) m.add(0);//
-        if (param&CM_E5) m.add(0);//
-        if (param&CM_MR) m.add(0);
-      }
-    }
-    if (bitPos==7) index = 0;
-    return result;
-  }
-  int inputs() {return inputCount;}
-};
-
-
 
 
 
@@ -7566,997 +7242,7 @@ int p(Mixer& m,int info,int val2=0){
  
 };
 
-#define CacheSize (1<<5)
 
-#if (CacheSize&(CacheSize-1)) || (CacheSize<8)
-  #error Cache size must be a power of 2 bigger than 4
-#endif
-//////////////////////////// exeModel /////////////////////////
-
-// Model x86 code.  The contexts are sparse containing only those
-// bits relevant to parsing (2 prefixes, opcode, and mod and r/m fields
-// of modR/M byte).
-
-// formats
-enum InstructionFormat {
-  // encoding mode
-  fNM = 0x0,      // no ModRM
-  fAM = 0x1,      // no ModRM, "address mode" (jumps or direct addresses)
-  fMR = 0x2,      // ModRM present
-  fMEXTRA = 0x3,  // ModRM present, includes extra bits for opcode
-  fMODE = 0x3,    // bitmask for mode
-
-  // no ModRM: size of immediate operand
-  fNI = 0x0,      // no immediate
-  fBI = 0x4,      // byte immediate
-  fWI = 0x8,      // word immediate
-  fDI = 0xc,      // dword immediate
-  fTYPE = 0xc,    // type mask
-
-  // address mode: type of address operand
-  fAD = 0x0,      // absolute address
-  fDA = 0x4,      // dword absolute jump target
-  fBR = 0x8,      // byte relative jump target
-  fDR = 0xc,      // dword relative jump target
-
-  // others
-  fERR = 0xf     // denotes invalid opcodes
-};
-
-// 1 byte opcodes
-const static U8 Table1[256] = {
-  // 0       1       2       3       4       5       6       7       8       9       a       b       c       d       e       f
-  fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI,fNM|fBI,fNM|fDI,fNM|fNI,fNM|fNI,fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI,fNM|fBI,fNM|fDI,fNM|fNI,fNM|fNI, // 0
-  fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI,fNM|fBI,fNM|fDI,fNM|fNI,fNM|fNI,fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI,fNM|fBI,fNM|fDI,fNM|fNI,fNM|fNI, // 1
-  fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI,fNM|fBI,fNM|fDI,fNM|fNI,fNM|fNI,fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI,fNM|fBI,fNM|fDI,fNM|fNI,fNM|fNI, // 2
-  fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI,fNM|fBI,fNM|fDI,fNM|fNI,fNM|fNI,fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI,fNM|fBI,fNM|fDI,fNM|fNI,fNM|fNI, // 3
-
-  fNM|fNI,fNM|fNI,fNM|fNI,fNM|fNI,fNM|fNI,fNM|fNI,fNM|fNI,fNM|fNI,fNM|fNI,fNM|fNI,fNM|fNI,fNM|fNI,fNM|fNI,fNM|fNI,fNM|fNI,fNM|fNI, // 4
-  fNM|fNI,fNM|fNI,fNM|fNI,fNM|fNI,fNM|fNI,fNM|fNI,fNM|fNI,fNM|fNI,fNM|fNI,fNM|fNI,fNM|fNI,fNM|fNI,fNM|fNI,fNM|fNI,fNM|fNI,fNM|fNI, // 5
-  fNM|fNI,fNM|fNI,fMR|fNI,fMR|fNI,fNM|fNI,fNM|fNI,fNM|fNI,fNM|fNI,fNM|fDI,fMR|fDI,fNM|fBI,fMR|fBI,fNM|fNI,fNM|fNI,fNM|fNI,fNM|fNI, // 6
-  fAM|fBR,fAM|fBR,fAM|fBR,fAM|fBR,fAM|fBR,fAM|fBR,fAM|fBR,fAM|fBR,fAM|fBR,fAM|fBR,fAM|fBR,fAM|fBR,fAM|fBR,fAM|fBR,fAM|fBR,fAM|fBR, // 7
-
-  fMR|fBI,fMR|fDI,fMR|fBI,fMR|fBI,fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI, // 8
-  fNM|fNI,fNM|fNI,fNM|fNI,fNM|fNI,fNM|fNI,fNM|fNI,fNM|fNI,fNM|fNI,fNM|fNI,fNM|fNI,fAM|fDA,fNM|fNI,fNM|fNI,fNM|fNI,fNM|fNI,fNM|fNI, // 9
-  fAM|fAD,fAM|fAD,fAM|fAD,fAM|fAD,fNM|fNI,fNM|fNI,fNM|fNI,fNM|fNI,fNM|fBI,fNM|fDI,fNM|fNI,fNM|fNI,fNM|fNI,fNM|fNI,fNM|fNI,fNM|fNI, // a
-  fNM|fBI,fNM|fBI,fNM|fBI,fNM|fBI,fNM|fBI,fNM|fBI,fNM|fBI,fNM|fBI,fNM|fDI,fNM|fDI,fNM|fDI,fNM|fDI,fNM|fDI,fNM|fDI,fNM|fDI,fNM|fDI, // b
-
-  fMR|fBI,fMR|fBI,fNM|fWI,fNM|fNI,fMR|fNI,fMR|fNI,fMR|fBI,fMR|fDI,fNM|fBI,fNM|fNI,fNM|fWI,fNM|fNI,fNM|fNI,fNM|fBI,fERR   ,fNM|fNI, // c
-  fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI,fNM|fBI,fNM|fBI,fNM|fNI,fNM|fNI,fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI, // d
-  fAM|fBR,fAM|fBR,fAM|fBR,fAM|fBR,fNM|fBI,fNM|fBI,fNM|fBI,fNM|fBI,fAM|fDR,fAM|fDR,fAM|fAD,fAM|fBR,fNM|fNI,fNM|fNI,fNM|fNI,fNM|fNI, // e
-  fNM|fNI,fERR   ,fNM|fNI,fNM|fNI,fNM|fNI,fNM|fNI,fMEXTRA,fMEXTRA,fNM|fNI,fNM|fNI,fNM|fNI,fNM|fNI,fNM|fNI,fNM|fNI,fMEXTRA,fMEXTRA, // f
-};
-
-// 2 byte opcodes
-const static U8 Table2[256] = {
-  // 0       1       2       3       4       5       6       7       8       9       a       b       c       d       e       f
-  fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fNM|fNI,fERR   ,fNM|fNI,fNM|fNI,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   , // 0
-  fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   , // 1
-  fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI,fERR   ,fERR   ,fERR   ,fERR   ,fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI, // 2
-  fNM|fNI,fNM|fNI,fNM|fNI,fNM|fNI,fNM|fNI,fNM|fNI,fERR   ,fNM|fNI,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   , // 3
-
-  fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI, // 4
-  fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI, // 5
-  fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI, // 6
-  fMR|fBI,fMR|fBI,fMR|fBI,fMR|fBI,fMR|fNI,fMR|fNI,fMR|fNI,fNM|fNI,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fMR|fNI,fMR|fNI, // 7
-
-  fAM|fDR,fAM|fDR,fAM|fDR,fAM|fDR,fAM|fDR,fAM|fDR,fAM|fDR,fAM|fDR,fAM|fDR,fAM|fDR,fAM|fDR,fAM|fDR,fAM|fDR,fAM|fDR,fAM|fDR,fAM|fDR, // 8
-  fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI, // 9
-  fNM|fNI,fNM|fNI,fNM|fNI,fMR|fNI,fMR|fBI,fMR|fNI,fMR|fNI,fMR|fNI,fERR   ,fERR   ,fERR   ,fMR|fNI,fMR|fBI,fMR|fNI,fERR   ,fMR|fNI, // a
-  fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI,fERR   ,fERR   ,fERR   ,fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI, // b
-
-  fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI,fNM|fNI,fNM|fNI,fNM|fNI,fNM|fNI,fNM|fNI,fNM|fNI,fNM|fNI,fNM|fNI, // c
-  fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI, // d
-  fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI, // e
-  fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI,fERR   , // f
-};
-
-// 3 byte opcodes $0F38XX
-const static U8 Table3_38[256] = {
-  // 0       1       2       3       4       5       6       7       8       9       a       b       c       d       e       f
-  fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI,fERR   ,fERR   ,fERR   ,fERR   , // 0
-  fMR|fNI,fERR   ,fERR   ,fERR   ,fMR|fNI,fMR|fNI,fERR   ,fMR|fNI,fERR   ,fERR   ,fERR   ,fERR   ,fMR|fNI,fMR|fNI,fMR|fNI,fERR   , // 1
-  fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI,fERR   ,fERR   ,fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI,fERR   ,fERR   ,fERR   ,fERR   , // 2
-  fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI,fERR   ,fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI, // 3
-  fMR|fNI,fMR|fNI,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   , // 4
-  fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   , // 5
-  fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   , // 6
-  fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   , // 7
-  fMR|fNI,fMR|fNI,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   , // 8
-  fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   , // 9
-  fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   , // a
-  fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   , // b
-  fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   , // c
-  fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   , // d
-  fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   , // e
-  fMR|fNI,fMR|fNI,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   , // f
-};
-
-// 3 byte opcodes $0F3AXX
-const static U8 Table3_3A[256] = {
-  // 0       1       2       3       4       5       6       7       8       9       a       b       c       d       e       f
-  fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fMR|fBI,fMR|fBI,fMR|fBI,fMR|fBI,fMR|fBI,fMR|fBI,fMR|fBI,fMR|fBI, // 0
-  fERR   ,fERR   ,fERR   ,fERR   ,fMR|fBI,fMR|fBI,fMR|fBI,fMR|fBI,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   , // 1
-  fMR|fBI,fMR|fBI,fMR|fBI,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   , // 2
-  fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   , // 3
-  fMR|fBI,fMR|fBI,fMR|fBI,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   , // 4
-  fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   , // 5
-  fMR|fBI,fMR|fBI,fMR|fBI,fMR|fBI,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   , // 6
-  fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   , // 7
-  fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   , // 8
-  fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   , // 9
-  fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   , // a
-  fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   , // b
-  fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   , // c
-  fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   , // d
-  fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   , // e
-  fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   , // f
-};
-
-// escape opcodes using ModRM byte to get more variants
-const static U8 TableX[32] = {
-  // 0       1       2       3       4       5       6       7
-  fMR|fBI,fERR   ,fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI, // escapes for 0xf6
-  fMR|fDI,fERR   ,fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI,fMR|fNI, // escapes for 0xf7
-  fMR|fNI,fMR|fNI,fERR   ,fERR   ,fERR   ,fERR   ,fERR   ,fERR   , // escapes for 0xfe
-  fMR|fNI,fMR|fNI,fMR|fNI,fERR   ,fMR|fNI,fERR   ,fMR|fNI,fERR   , // escapes for 0xff
-};
-
-const static U8 InvalidX64Ops[19] = {0x06, 0x07, 0x16, 0x17, 0x1E, 0x1F, 0x27, 0x2F, 0x37, 0x3F, 0x60, 0x61, 0x62, 0x82, 0x9A, 0xD4, 0xD5, 0xD6, 0xEA,};
-const static U8 X64Prefixes[8] = {0x26, 0x2E, 0x36, 0x3E, 0x9B, 0xF0, 0xF2, 0xF3,};
-enum InstructionCategory {
-  OP_INVALID              =  0,
-  OP_PREFIX_SEGREG        =  1,
-  OP_PREFIX               =  2,
-  OP_PREFIX_X87FPU        =  3,
-  OP_GEN_DATAMOV          =  4,
-  OP_GEN_STACK            =  5,
-  OP_GEN_CONVERSION       =  6,
-  OP_GEN_ARITH_DECIMAL    =  7,
-  OP_GEN_ARITH_BINARY     =  8,
-  OP_GEN_LOGICAL          =  9,
-  OP_GEN_SHF_ROT          = 10,
-  OP_GEN_BIT              = 11,
-  OP_GEN_BRANCH           = 12,
-  OP_GEN_BRANCH_COND      = 13,
-  OP_GEN_BREAK            = 14,
-  OP_GEN_STRING           = 15,
-  OP_GEN_INOUT            = 16,
-  OP_GEN_FLAG_CONTROL     = 17,
-  OP_GEN_SEGREG           = 18,
-  OP_GEN_CONTROL          = 19,
-  OP_SYSTEM               = 20,
-  OP_X87_DATAMOV          = 21,
-  OP_X87_ARITH            = 22,
-  OP_X87_COMPARISON       = 23,
-  OP_X87_TRANSCENDENTAL   = 24,
-  OP_X87_LOAD_CONSTANT    = 25,
-  OP_X87_CONTROL          = 26,
-  OP_X87_CONVERSION       = 27,
-  OP_STATE_MANAGEMENT     = 28,
-  OP_MMX                  = 29,
-  OP_SSE                  = 30,
-  OP_SSE_DATAMOV          = 31
-};
-
-const static U8 TypeOp1[256] = {
-  OP_GEN_ARITH_BINARY  , OP_GEN_ARITH_BINARY  , OP_GEN_ARITH_BINARY  , OP_GEN_ARITH_BINARY  , //03
-  OP_GEN_ARITH_BINARY  , OP_GEN_ARITH_BINARY  , OP_GEN_STACK         , OP_GEN_STACK         , //07
-  OP_GEN_LOGICAL       , OP_GEN_LOGICAL       , OP_GEN_LOGICAL       , OP_GEN_LOGICAL       , //0B
-  OP_GEN_LOGICAL       , OP_GEN_LOGICAL       , OP_GEN_STACK         , OP_PREFIX            , //0F
-  OP_GEN_ARITH_BINARY  , OP_GEN_ARITH_BINARY  , OP_GEN_ARITH_BINARY  , OP_GEN_ARITH_BINARY  , //13
-  OP_GEN_ARITH_BINARY  , OP_GEN_ARITH_BINARY  , OP_GEN_STACK         , OP_GEN_STACK         , //17
-  OP_GEN_ARITH_BINARY  , OP_GEN_ARITH_BINARY  , OP_GEN_ARITH_BINARY  , OP_GEN_ARITH_BINARY  , //1B
-  OP_GEN_ARITH_BINARY  , OP_GEN_ARITH_BINARY  , OP_GEN_STACK         , OP_GEN_STACK         , //1F
-  OP_GEN_LOGICAL       , OP_GEN_LOGICAL       , OP_GEN_LOGICAL       , OP_GEN_LOGICAL       , //23
-  OP_GEN_LOGICAL       , OP_GEN_LOGICAL       , OP_PREFIX_SEGREG     , OP_GEN_ARITH_DECIMAL , //27
-  OP_GEN_ARITH_BINARY  , OP_GEN_ARITH_BINARY  , OP_GEN_ARITH_BINARY  , OP_GEN_ARITH_BINARY  , //2B
-  OP_GEN_ARITH_BINARY  , OP_GEN_ARITH_BINARY  , OP_PREFIX_SEGREG     , OP_GEN_ARITH_DECIMAL , //2F
-  OP_GEN_LOGICAL       , OP_GEN_LOGICAL       , OP_GEN_LOGICAL       , OP_GEN_LOGICAL       , //33
-  OP_GEN_LOGICAL       , OP_GEN_LOGICAL       , OP_PREFIX_SEGREG     , OP_GEN_ARITH_DECIMAL , //37
-  OP_GEN_ARITH_BINARY  , OP_GEN_ARITH_BINARY  , OP_GEN_ARITH_BINARY  , OP_GEN_ARITH_BINARY  , //3B
-  OP_GEN_ARITH_BINARY  , OP_GEN_ARITH_BINARY  , OP_PREFIX_SEGREG     , OP_GEN_ARITH_DECIMAL , //3F
-  OP_GEN_ARITH_BINARY  , OP_GEN_ARITH_BINARY  , OP_GEN_ARITH_BINARY  , OP_GEN_ARITH_BINARY  , //43
-  OP_GEN_ARITH_BINARY  , OP_GEN_ARITH_BINARY  , OP_GEN_ARITH_BINARY  , OP_GEN_ARITH_BINARY  , //47
-  OP_GEN_ARITH_BINARY  , OP_GEN_ARITH_BINARY  , OP_GEN_ARITH_BINARY  , OP_GEN_ARITH_BINARY  , //4B
-  OP_GEN_ARITH_BINARY  , OP_GEN_ARITH_BINARY  , OP_GEN_ARITH_BINARY  , OP_GEN_ARITH_BINARY  , //4F
-  OP_GEN_STACK         , OP_GEN_STACK         , OP_GEN_STACK         , OP_GEN_STACK         , //53
-  OP_GEN_STACK         , OP_GEN_STACK         , OP_GEN_STACK         , OP_GEN_STACK         , //57
-  OP_GEN_STACK         , OP_GEN_STACK         , OP_GEN_STACK         , OP_GEN_STACK         , //5B
-  OP_GEN_STACK         , OP_GEN_STACK         , OP_GEN_STACK         , OP_GEN_STACK         , //5F
-  OP_GEN_STACK         , OP_GEN_STACK         , OP_GEN_BREAK         , OP_GEN_CONVERSION    , //63
-  OP_PREFIX_SEGREG     , OP_PREFIX_SEGREG     , OP_PREFIX            , OP_PREFIX            , //67
-  OP_GEN_STACK         , OP_GEN_ARITH_BINARY  , OP_GEN_STACK         , OP_GEN_ARITH_BINARY  , //6B
-  OP_GEN_INOUT         , OP_GEN_INOUT         , OP_GEN_INOUT         , OP_GEN_INOUT         , //6F
-  OP_GEN_BRANCH_COND   , OP_GEN_BRANCH_COND   , OP_GEN_BRANCH_COND   , OP_GEN_BRANCH_COND   , //73
-  OP_GEN_BRANCH_COND   , OP_GEN_BRANCH_COND   , OP_GEN_BRANCH_COND   , OP_GEN_BRANCH_COND   , //77
-  OP_GEN_BRANCH_COND   , OP_GEN_BRANCH_COND   , OP_GEN_BRANCH_COND   , OP_GEN_BRANCH_COND   , //7B
-  OP_GEN_BRANCH_COND   , OP_GEN_BRANCH_COND   , OP_GEN_BRANCH_COND   , OP_GEN_BRANCH_COND   , //7F
-  OP_GEN_ARITH_BINARY  , OP_GEN_ARITH_BINARY  , OP_GEN_ARITH_BINARY  , OP_GEN_ARITH_BINARY  , //83
-  OP_GEN_ARITH_BINARY  , OP_GEN_ARITH_BINARY  , OP_GEN_DATAMOV       , OP_GEN_DATAMOV       , //87
-  OP_GEN_DATAMOV       , OP_GEN_DATAMOV       , OP_GEN_DATAMOV       , OP_GEN_DATAMOV       , //8B
-  OP_GEN_DATAMOV       , OP_GEN_DATAMOV       , OP_GEN_DATAMOV       , OP_GEN_STACK         , //8F
-  OP_GEN_DATAMOV       , OP_GEN_DATAMOV       , OP_GEN_DATAMOV       , OP_GEN_DATAMOV       , //93
-  OP_GEN_DATAMOV       , OP_GEN_DATAMOV       , OP_GEN_DATAMOV       , OP_GEN_DATAMOV       , //97
-  OP_GEN_CONVERSION    , OP_GEN_CONVERSION    , OP_GEN_BRANCH        , OP_PREFIX_X87FPU     , //9B
-  OP_GEN_STACK         , OP_GEN_STACK         , OP_GEN_DATAMOV       , OP_GEN_DATAMOV       , //9F
-  OP_GEN_DATAMOV       , OP_GEN_DATAMOV       , OP_GEN_DATAMOV       , OP_GEN_DATAMOV       , //A3
-  OP_GEN_STRING        , OP_GEN_STRING        , OP_GEN_STRING        , OP_GEN_STRING        , //A7
-  OP_GEN_LOGICAL       , OP_GEN_LOGICAL       , OP_GEN_STRING        , OP_GEN_STRING        , //AB
-  OP_GEN_STRING        , OP_GEN_STRING        , OP_GEN_STRING        , OP_GEN_STRING        , //AF
-  OP_GEN_DATAMOV       , OP_GEN_DATAMOV       , OP_GEN_DATAMOV       , OP_GEN_DATAMOV       , //B3
-  OP_GEN_DATAMOV       , OP_GEN_DATAMOV       , OP_GEN_DATAMOV       , OP_GEN_DATAMOV       , //B7
-  OP_GEN_DATAMOV       , OP_GEN_DATAMOV       , OP_GEN_DATAMOV       , OP_GEN_DATAMOV       , //BB
-  OP_GEN_DATAMOV       , OP_GEN_DATAMOV       , OP_GEN_DATAMOV       , OP_GEN_DATAMOV       , //BF
-  OP_GEN_SHF_ROT       , OP_GEN_SHF_ROT       , OP_GEN_BRANCH        , OP_GEN_BRANCH        , //C3
-  OP_GEN_DATAMOV       , OP_GEN_DATAMOV       , OP_GEN_DATAMOV       , OP_GEN_DATAMOV       , //C7
-  OP_GEN_STACK         , OP_GEN_STACK         , OP_GEN_BRANCH        , OP_GEN_BRANCH        , //CB
-  OP_GEN_BREAK         , OP_GEN_BREAK         , OP_GEN_BREAK         , OP_GEN_BREAK         , //CF
-  OP_GEN_SHF_ROT       , OP_GEN_SHF_ROT       , OP_GEN_SHF_ROT       , OP_GEN_SHF_ROT       , //D3
-  OP_GEN_ARITH_DECIMAL , OP_GEN_ARITH_DECIMAL , OP_GEN_DATAMOV       , OP_GEN_DATAMOV       , //D7
-  OP_X87_ARITH         , OP_X87_DATAMOV       , OP_X87_ARITH         , OP_X87_DATAMOV       , //DB
-  OP_X87_ARITH         , OP_X87_DATAMOV       , OP_X87_ARITH         , OP_X87_DATAMOV       , //DF
-  OP_GEN_BRANCH_COND   , OP_GEN_BRANCH_COND   , OP_GEN_BRANCH_COND   , OP_GEN_BRANCH_COND   , //E3
-  OP_GEN_INOUT         , OP_GEN_INOUT         , OP_GEN_INOUT         , OP_GEN_INOUT         , //E7
-  OP_GEN_BRANCH        , OP_GEN_BRANCH        , OP_GEN_BRANCH        , OP_GEN_BRANCH        , //EB
-  OP_GEN_INOUT         , OP_GEN_INOUT         , OP_GEN_INOUT         , OP_GEN_INOUT         , //EF
-  OP_PREFIX            , OP_GEN_BREAK         , OP_PREFIX            , OP_PREFIX            , //F3
-  OP_SYSTEM            , OP_GEN_FLAG_CONTROL  , OP_GEN_ARITH_BINARY  , OP_GEN_ARITH_BINARY  , //F7
-  OP_GEN_FLAG_CONTROL  , OP_GEN_FLAG_CONTROL  , OP_GEN_FLAG_CONTROL  , OP_GEN_FLAG_CONTROL  , //FB
-  OP_GEN_FLAG_CONTROL  , OP_GEN_FLAG_CONTROL  , OP_GEN_ARITH_BINARY  , OP_GEN_BRANCH         //FF
-};
-
-const static U8 TypeOp2[256] = {
-  OP_SYSTEM            , OP_SYSTEM            , OP_SYSTEM            , OP_SYSTEM            , //03
-  OP_INVALID           , OP_SYSTEM            , OP_SYSTEM            , OP_SYSTEM            , //07
-  OP_SYSTEM            , OP_SYSTEM            , OP_INVALID           , OP_GEN_CONTROL       , //0B
-  OP_INVALID           , OP_GEN_CONTROL       , OP_INVALID           , OP_INVALID           , //0F
-  OP_SSE_DATAMOV       , OP_SSE_DATAMOV       , OP_SSE_DATAMOV       , OP_SSE_DATAMOV       , //13
-  OP_SSE               , OP_SSE               , OP_SSE_DATAMOV       , OP_SSE_DATAMOV       , //17
-  OP_SSE               , OP_GEN_CONTROL       , OP_GEN_CONTROL       , OP_GEN_CONTROL       , //1B
-  OP_GEN_CONTROL       , OP_GEN_CONTROL       , OP_GEN_CONTROL       , OP_GEN_CONTROL       , //1F
-  OP_SYSTEM            , OP_SYSTEM            , OP_SYSTEM            , OP_SYSTEM            , //23
-  OP_SYSTEM            , OP_INVALID           , OP_SYSTEM            , OP_INVALID           , //27
-  OP_SSE_DATAMOV       , OP_SSE_DATAMOV       , OP_SSE               , OP_SSE               , //2B
-  OP_SSE               , OP_SSE               , OP_SSE               , OP_SSE               , //2F
-  OP_SYSTEM            , OP_SYSTEM            , OP_SYSTEM            , OP_SYSTEM            , //33
-  OP_SYSTEM            , OP_SYSTEM            , OP_INVALID           , OP_INVALID           , //37
-  OP_PREFIX            , OP_INVALID           , OP_PREFIX            , OP_INVALID           , //3B
-  OP_INVALID           , OP_INVALID           , OP_INVALID           , OP_INVALID           , //3F
-  OP_GEN_DATAMOV       , OP_GEN_DATAMOV       , OP_GEN_DATAMOV       , OP_GEN_DATAMOV       , //43
-  OP_GEN_DATAMOV       , OP_GEN_DATAMOV       , OP_GEN_DATAMOV       , OP_GEN_DATAMOV       , //47
-  OP_GEN_DATAMOV       , OP_GEN_DATAMOV       , OP_GEN_DATAMOV       , OP_GEN_DATAMOV       , //4B
-  OP_GEN_DATAMOV       , OP_GEN_DATAMOV       , OP_GEN_DATAMOV       , OP_GEN_DATAMOV       , //4F
-  OP_SSE_DATAMOV       , OP_SSE               , OP_SSE               , OP_SSE               , //53
-  OP_SSE               , OP_SSE               , OP_SSE               , OP_SSE               , //57
-  OP_SSE               , OP_SSE               , OP_SSE               , OP_SSE               , //5B
-  OP_SSE               , OP_SSE               , OP_SSE               , OP_SSE               , //5F
-  OP_MMX               , OP_MMX               , OP_MMX               , OP_MMX               , //63
-  OP_MMX               , OP_MMX               , OP_MMX               , OP_MMX               , //67
-  OP_MMX               , OP_MMX               , OP_MMX               , OP_MMX               , //6B
-  OP_INVALID           , OP_INVALID           , OP_MMX               , OP_MMX               , //6F
-  OP_SSE               , OP_MMX               , OP_MMX               , OP_MMX               , //73
-  OP_MMX               , OP_MMX               , OP_MMX               , OP_MMX               , //77
-  OP_INVALID           , OP_INVALID           , OP_INVALID           , OP_INVALID           , //7B
-  OP_INVALID           , OP_INVALID           , OP_MMX               , OP_MMX               , //7F
-  OP_GEN_BRANCH_COND   , OP_GEN_BRANCH_COND   , OP_GEN_BRANCH_COND   , OP_GEN_BRANCH_COND   , //83
-  OP_GEN_BRANCH_COND   , OP_GEN_BRANCH_COND   , OP_GEN_BRANCH_COND   , OP_GEN_BRANCH_COND   , //87
-  OP_GEN_BRANCH_COND   , OP_GEN_BRANCH_COND   , OP_GEN_BRANCH_COND   , OP_GEN_BRANCH_COND   , //8B
-  OP_GEN_BRANCH_COND   , OP_GEN_BRANCH_COND   , OP_GEN_BRANCH_COND   , OP_GEN_BRANCH_COND   , //8F
-  OP_GEN_DATAMOV       , OP_GEN_DATAMOV       , OP_GEN_DATAMOV       , OP_GEN_DATAMOV       , //93
-  OP_GEN_DATAMOV       , OP_GEN_DATAMOV       , OP_GEN_DATAMOV       , OP_GEN_DATAMOV       , //97
-  OP_GEN_DATAMOV       , OP_GEN_DATAMOV       , OP_GEN_DATAMOV       , OP_GEN_DATAMOV       , //9B
-  OP_GEN_DATAMOV       , OP_GEN_DATAMOV       , OP_GEN_DATAMOV       , OP_GEN_DATAMOV       , //9F
-  OP_GEN_STACK         , OP_GEN_STACK         , OP_GEN_CONTROL       , OP_GEN_BIT           , //A3
-  OP_GEN_SHF_ROT       , OP_GEN_SHF_ROT       , OP_INVALID           , OP_INVALID           , //A7
-  OP_GEN_STACK         , OP_GEN_STACK         , OP_SYSTEM            , OP_GEN_BIT           , //AB
-  OP_GEN_SHF_ROT       , OP_GEN_SHF_ROT       , OP_STATE_MANAGEMENT  , OP_GEN_ARITH_BINARY  , //AF
-  OP_GEN_DATAMOV       , OP_GEN_DATAMOV       , OP_GEN_DATAMOV       , OP_GEN_BIT           , //B3
-  OP_GEN_DATAMOV       , OP_GEN_DATAMOV       , OP_GEN_CONVERSION    , OP_GEN_CONVERSION    , //B7
-  OP_INVALID           , OP_GEN_CONTROL       , OP_GEN_BIT           , OP_GEN_BIT           , //BB
-  OP_GEN_BIT           , OP_GEN_BIT           , OP_GEN_CONVERSION    , OP_GEN_CONVERSION    , //BF
-  OP_GEN_DATAMOV       , OP_GEN_DATAMOV       , OP_SSE               , OP_SSE               , //C3
-  OP_SSE               , OP_SSE               , OP_SSE               , OP_GEN_DATAMOV       , //C7
-  OP_GEN_DATAMOV       , OP_GEN_DATAMOV       , OP_GEN_DATAMOV       , OP_GEN_DATAMOV       , //CB
-  OP_GEN_DATAMOV       , OP_GEN_DATAMOV       , OP_GEN_DATAMOV       , OP_GEN_DATAMOV       , //CF
-  OP_INVALID           , OP_MMX               , OP_MMX               , OP_MMX               , //D3
-  OP_SSE               , OP_MMX               , OP_INVALID           , OP_SSE               , //D7
-  OP_MMX               , OP_MMX               , OP_SSE               , OP_MMX               , //DB
-  OP_MMX               , OP_MMX               , OP_SSE               , OP_MMX               , //DF
-  OP_SSE               , OP_MMX               , OP_SSE               , OP_MMX               , //E3
-  OP_SSE               , OP_MMX               , OP_INVALID           , OP_SSE               , //E7
-  OP_MMX               , OP_MMX               , OP_SSE               , OP_MMX               , //EB
-  OP_MMX               , OP_MMX               , OP_SSE               , OP_MMX               , //EF
-  OP_INVALID           , OP_MMX               , OP_MMX               , OP_MMX               , //F3
-  OP_SSE               , OP_MMX               , OP_SSE               , OP_SSE               , //F7
-  OP_MMX               , OP_MMX               , OP_MMX               , OP_SSE               , //FB
-  OP_MMX               , OP_MMX               , OP_MMX               , OP_INVALID            //FF
-};
-
-const static U8 TypeOp3_38[256] = {
-  OP_SSE               , OP_SSE               , OP_SSE               , OP_SSE               , //03
-  OP_SSE               , OP_SSE               , OP_SSE               , OP_SSE               , //07
-  OP_SSE               , OP_SSE               , OP_SSE               , OP_SSE               , //0B
-  OP_INVALID           , OP_INVALID           , OP_INVALID           , OP_INVALID           , //0F
-  OP_SSE               , OP_INVALID           , OP_INVALID           , OP_INVALID           , //13
-  OP_SSE               , OP_SSE               , OP_INVALID           , OP_SSE               , //17
-  OP_INVALID           , OP_INVALID           , OP_INVALID           , OP_INVALID           , //1B
-  OP_SSE               , OP_SSE               , OP_SSE               , OP_INVALID           , //1F
-  OP_SSE               , OP_SSE               , OP_SSE               , OP_SSE               , //23
-  OP_SSE               , OP_SSE               , OP_INVALID           , OP_INVALID           , //27
-  OP_SSE               , OP_SSE               , OP_SSE               , OP_SSE               , //2B
-  OP_INVALID           , OP_INVALID           , OP_INVALID           , OP_INVALID           , //2F
-  OP_SSE               , OP_SSE               , OP_SSE               , OP_SSE               , //33
-  OP_SSE               , OP_SSE               , OP_INVALID           , OP_SSE               , //37
-  OP_SSE               , OP_SSE               , OP_SSE               , OP_SSE               , //3B
-  OP_SSE               , OP_SSE               , OP_SSE               , OP_SSE               , //3F
-  OP_SSE               , OP_SSE               , OP_INVALID           , OP_INVALID           , //43
-  OP_INVALID           , OP_INVALID           , OP_INVALID           , OP_INVALID           , //47
-  OP_INVALID           , OP_INVALID           , OP_INVALID           , OP_INVALID           , //4B
-  OP_INVALID           , OP_INVALID           , OP_INVALID           , OP_INVALID           , //4F
-  OP_INVALID           , OP_INVALID           , OP_INVALID           , OP_INVALID           , //53
-  OP_INVALID           , OP_INVALID           , OP_INVALID           , OP_INVALID           , //57
-  OP_INVALID           , OP_INVALID           , OP_INVALID           , OP_INVALID           , //5B
-  OP_INVALID           , OP_INVALID           , OP_INVALID           , OP_INVALID           , //5F
-  OP_INVALID           , OP_INVALID           , OP_INVALID           , OP_INVALID           , //63
-  OP_INVALID           , OP_INVALID           , OP_INVALID           , OP_INVALID           , //67
-  OP_INVALID           , OP_INVALID           , OP_INVALID           , OP_INVALID           , //6B
-  OP_INVALID           , OP_INVALID           , OP_INVALID           , OP_INVALID           , //6F
-  OP_INVALID           , OP_INVALID           , OP_INVALID           , OP_INVALID           , //73
-  OP_INVALID           , OP_INVALID           , OP_INVALID           , OP_INVALID           , //77
-  OP_INVALID           , OP_INVALID           , OP_INVALID           , OP_INVALID           , //7B
-  OP_INVALID           , OP_INVALID           , OP_INVALID           , OP_INVALID           , //7F
-  OP_INVALID           , OP_INVALID           , OP_INVALID           , OP_INVALID           , //83
-  OP_INVALID           , OP_INVALID           , OP_INVALID           , OP_INVALID           , //87
-  OP_INVALID           , OP_INVALID           , OP_INVALID           , OP_INVALID           , //8B
-  OP_INVALID           , OP_INVALID           , OP_INVALID           , OP_INVALID           , //8F
-  OP_INVALID           , OP_INVALID           , OP_INVALID           , OP_INVALID           , //93
-  OP_INVALID           , OP_INVALID           , OP_INVALID           , OP_INVALID           , //97
-  OP_INVALID           , OP_INVALID           , OP_INVALID           , OP_INVALID           , //9B
-  OP_INVALID           , OP_INVALID           , OP_INVALID           , OP_INVALID           , //9F
-  OP_INVALID           , OP_INVALID           , OP_INVALID           , OP_INVALID           , //A3
-  OP_INVALID           , OP_INVALID           , OP_INVALID           , OP_INVALID           , //A7
-  OP_INVALID           , OP_INVALID           , OP_INVALID           , OP_INVALID           , //AB
-  OP_INVALID           , OP_INVALID           , OP_INVALID           , OP_INVALID           , //AF
-  OP_INVALID           , OP_INVALID           , OP_INVALID           , OP_INVALID           , //B3
-  OP_INVALID           , OP_INVALID           , OP_INVALID           , OP_INVALID           , //B7
-  OP_INVALID           , OP_INVALID           , OP_INVALID           , OP_INVALID           , //BB
-  OP_INVALID           , OP_INVALID           , OP_INVALID           , OP_INVALID           , //BF
-  OP_INVALID           , OP_INVALID           , OP_INVALID           , OP_INVALID           , //C3
-  OP_INVALID           , OP_INVALID           , OP_INVALID           , OP_INVALID           , //C7
-  OP_INVALID           , OP_INVALID           , OP_INVALID           , OP_INVALID           , //CB
-  OP_INVALID           , OP_INVALID           , OP_INVALID           , OP_INVALID           , //CF
-  OP_INVALID           , OP_INVALID           , OP_INVALID           , OP_INVALID           , //D3
-  OP_INVALID           , OP_INVALID           , OP_INVALID           , OP_INVALID           , //D7
-  OP_INVALID           , OP_INVALID           , OP_INVALID           , OP_INVALID           , //DB
-  OP_INVALID           , OP_INVALID           , OP_INVALID           , OP_INVALID           , //DF
-  OP_INVALID           , OP_INVALID           , OP_INVALID           , OP_INVALID           , //E3
-  OP_INVALID           , OP_INVALID           , OP_INVALID           , OP_INVALID           , //E7
-  OP_INVALID           , OP_INVALID           , OP_INVALID           , OP_INVALID           , //EB
-  OP_INVALID           , OP_INVALID           , OP_INVALID           , OP_INVALID           , //EF
-  OP_GEN_DATAMOV       , OP_GEN_DATAMOV       , OP_INVALID           , OP_INVALID           , //F3
-  OP_INVALID           , OP_INVALID           , OP_INVALID           , OP_INVALID           , //F7
-  OP_INVALID           , OP_INVALID           , OP_INVALID           , OP_INVALID           , //FB
-  OP_INVALID           , OP_INVALID           , OP_INVALID           , OP_INVALID            //FF
-};
-
-const static U8 TypeOp3_3A[256] = {
-  OP_INVALID           , OP_INVALID           , OP_INVALID           , OP_INVALID           , //03
-  OP_INVALID           , OP_INVALID           , OP_INVALID           , OP_INVALID           , //07
-  OP_SSE               , OP_SSE               , OP_SSE               , OP_SSE               , //0B
-  OP_SSE               , OP_SSE               , OP_SSE               , OP_SSE               , //0F
-  OP_INVALID           , OP_INVALID           , OP_INVALID           , OP_INVALID           , //13
-  OP_SSE               , OP_SSE               , OP_SSE               , OP_SSE               , //17
-  OP_INVALID           , OP_INVALID           , OP_INVALID           , OP_INVALID           , //1B
-  OP_INVALID           , OP_INVALID           , OP_INVALID           , OP_INVALID           , //1F
-  OP_SSE               , OP_SSE               , OP_SSE               , OP_INVALID           , //23
-  OP_INVALID           , OP_INVALID           , OP_INVALID           , OP_INVALID           , //27
-  OP_INVALID           , OP_INVALID           , OP_INVALID           , OP_INVALID           , //2B
-  OP_INVALID           , OP_INVALID           , OP_INVALID           , OP_INVALID           , //2F
-  OP_INVALID           , OP_INVALID           , OP_INVALID           , OP_INVALID           , //33
-  OP_INVALID           , OP_INVALID           , OP_INVALID           , OP_INVALID           , //37
-  OP_INVALID           , OP_INVALID           , OP_INVALID           , OP_INVALID           , //3B
-  OP_INVALID           , OP_INVALID           , OP_INVALID           , OP_INVALID           , //3F
-  OP_SSE               , OP_SSE               , OP_SSE               , OP_INVALID           , //43
-  OP_INVALID           , OP_INVALID           , OP_INVALID           , OP_INVALID           , //47
-  OP_INVALID           , OP_INVALID           , OP_INVALID           , OP_INVALID           , //4B
-  OP_INVALID           , OP_INVALID           , OP_INVALID           , OP_INVALID           , //4F
-  OP_INVALID           , OP_INVALID           , OP_INVALID           , OP_INVALID           , //53
-  OP_INVALID           , OP_INVALID           , OP_INVALID           , OP_INVALID           , //57
-  OP_INVALID           , OP_INVALID           , OP_INVALID           , OP_INVALID           , //5B
-  OP_INVALID           , OP_INVALID           , OP_INVALID           , OP_INVALID           , //5F
-  OP_SSE               , OP_SSE               , OP_SSE               , OP_SSE               , //63
-  OP_INVALID           , OP_INVALID           , OP_INVALID           , OP_INVALID           , //67
-  OP_INVALID           , OP_INVALID           , OP_INVALID           , OP_INVALID           , //6B
-  OP_INVALID           , OP_INVALID           , OP_INVALID           , OP_INVALID           , //6F
-  OP_INVALID           , OP_INVALID           , OP_INVALID           , OP_INVALID           , //73
-  OP_INVALID           , OP_INVALID           , OP_INVALID           , OP_INVALID           , //77
-  OP_INVALID           , OP_INVALID           , OP_INVALID           , OP_INVALID           , //7B
-  OP_INVALID           , OP_INVALID           , OP_INVALID           , OP_INVALID           , //7F
-  OP_INVALID           , OP_INVALID           , OP_INVALID           , OP_INVALID           , //83
-  OP_INVALID           , OP_INVALID           , OP_INVALID           , OP_INVALID           , //87
-  OP_INVALID           , OP_INVALID           , OP_INVALID           , OP_INVALID           , //8B
-  OP_INVALID           , OP_INVALID           , OP_INVALID           , OP_INVALID           , //8F
-  OP_INVALID           , OP_INVALID           , OP_INVALID           , OP_INVALID           , //93
-  OP_INVALID           , OP_INVALID           , OP_INVALID           , OP_INVALID           , //97
-  OP_INVALID           , OP_INVALID           , OP_INVALID           , OP_INVALID           , //9B
-  OP_INVALID           , OP_INVALID           , OP_INVALID           , OP_INVALID           , //9F
-  OP_INVALID           , OP_INVALID           , OP_INVALID           , OP_INVALID           , //A3
-  OP_INVALID           , OP_INVALID           , OP_INVALID           , OP_INVALID           , //A7
-  OP_INVALID           , OP_INVALID           , OP_INVALID           , OP_INVALID           , //AB
-  OP_INVALID           , OP_INVALID           , OP_INVALID           , OP_INVALID           , //AF
-  OP_INVALID           , OP_INVALID           , OP_INVALID           , OP_INVALID           , //B3
-  OP_INVALID           , OP_INVALID           , OP_INVALID           , OP_INVALID           , //B7
-  OP_INVALID           , OP_INVALID           , OP_INVALID           , OP_INVALID           , //BB
-  OP_INVALID           , OP_INVALID           , OP_INVALID           , OP_INVALID           , //BF
-  OP_INVALID           , OP_INVALID           , OP_INVALID           , OP_INVALID           , //C3
-  OP_INVALID           , OP_INVALID           , OP_INVALID           , OP_INVALID           , //C7
-  OP_INVALID           , OP_INVALID           , OP_INVALID           , OP_INVALID           , //CB
-  OP_INVALID           , OP_INVALID           , OP_INVALID           , OP_INVALID           , //CF
-  OP_INVALID           , OP_INVALID           , OP_INVALID           , OP_INVALID           , //D3
-  OP_INVALID           , OP_INVALID           , OP_INVALID           , OP_INVALID           , //D7
-  OP_INVALID           , OP_INVALID           , OP_INVALID           , OP_INVALID           , //DB
-  OP_INVALID           , OP_INVALID           , OP_INVALID           , OP_INVALID           , //DF
-  OP_INVALID           , OP_INVALID           , OP_INVALID           , OP_INVALID           , //E3
-  OP_INVALID           , OP_INVALID           , OP_INVALID           , OP_INVALID           , //E7
-  OP_INVALID           , OP_INVALID           , OP_INVALID           , OP_INVALID           , //EB
-  OP_INVALID           , OP_INVALID           , OP_INVALID           , OP_INVALID           , //EF
-  OP_INVALID           , OP_INVALID           , OP_INVALID           , OP_INVALID           , //F3
-  OP_INVALID           , OP_INVALID           , OP_INVALID           , OP_INVALID           , //F7
-  OP_INVALID           , OP_INVALID           , OP_INVALID           , OP_INVALID           , //FB
-  OP_INVALID           , OP_INVALID           , OP_INVALID           , OP_INVALID            //FF
-};
-
-const static U8 TypeOpX[32] = {
-  // escapes for F6
-  OP_GEN_LOGICAL       , OP_GEN_LOGICAL       , OP_GEN_LOGICAL       , OP_GEN_ARITH_BINARY  ,
-  OP_GEN_ARITH_BINARY  , OP_GEN_ARITH_BINARY  , OP_GEN_ARITH_BINARY  , OP_GEN_ARITH_BINARY  ,
-  // escapes for F7
-  OP_GEN_LOGICAL       , OP_GEN_LOGICAL       , OP_GEN_LOGICAL       , OP_GEN_ARITH_BINARY  ,
-  OP_GEN_ARITH_BINARY  , OP_GEN_ARITH_BINARY  , OP_GEN_ARITH_BINARY  , OP_GEN_ARITH_BINARY  ,
-  // escapes for FE
-  OP_GEN_ARITH_BINARY  , OP_GEN_ARITH_BINARY  , OP_INVALID           , OP_INVALID           ,
-  OP_INVALID           , OP_INVALID           , OP_INVALID           , OP_INVALID           ,
-  // escapes for FF
-  OP_GEN_ARITH_BINARY  , OP_GEN_ARITH_BINARY  , OP_GEN_BRANCH        , OP_GEN_BRANCH        ,
-  OP_GEN_BRANCH        , OP_GEN_BRANCH        , OP_GEN_STACK         , OP_INVALID           
-};
-enum Prefixes {
-  ES_OVERRIDE = 0x26,
-  CS_OVERRIDE = 0x2E,
-  SS_OVERRIDE = 0x36,
-  DS_OVERRIDE = 0x3E,
-  FS_OVERRIDE = 0x64,
-  GS_OVERRIDE = 0x65,
-  AD_OVERRIDE = 0x67,
-  WAIT_FPU    = 0x9B,
-  LOCK        = 0xF0,
-  REP_N_STR   = 0xF2,
-  REP_STR     = 0xF3
-};
-
-enum Opcodes {
-  // 1-byte opcodes of special interest (for one reason or another)
-  OP_2BYTE  = 0x0f,     // start of 2-byte opcode
-  OP_OSIZE  = 0x66,     // operand size prefix
-  OP_CALLF  = 0x9a,
-  OP_RETNI  = 0xc2,     // ret near+immediate
-  OP_RETN   = 0xc3,
-  OP_ENTER  = 0xc8,
-  OP_INT3   = 0xcc,
-  OP_INTO   = 0xce,
-  OP_CALLN  = 0xe8,
-  OP_JMPF   = 0xea,
-  OP_ICEBP  = 0xf1
-};
-
-enum ExeState {
-  Start             =  0,
-  Pref_Op_Size      =  1,
-  Pref_MultiByte_Op =  2,
-  ParseFlags        =  3,
-  ExtraFlags        =  4,
-  ReadModRM         =  5,
-  Read_OP3_38       =  6,
-  Read_OP3_3A       =  7,
-  ReadSIB           =  8,
-  Read8             =  9,
-  Read16            = 10,
-  Read32            = 11,
-  Read8_ModRM       = 12,
-  Read16_f          = 13,
-  Read32_ModRM      = 14,
-  Error             = 15
-};
-
-struct OpCache {
-  U32 Op[CacheSize];
-  U32 Index;
-};
-
-struct Instruction {
-  U32 Data;
-  U8 Prefix, Code, ModRM, SIB, REX, Flags, BytesRead, Size, Category;
-  bool MustCheckREX, Decoding, o16, imm8;
-};
-
-#define CodeShift            3
-#define CodeMask             (0xFF<<CodeShift)
-#define ClearCodeMask        ((-1)^CodeMask)
-#define PrefixMask           ((1<<CodeShift)-1)
-#define OperandSizeOverride  (0x01<<(8+CodeShift))
-#define MultiByteOpcode      (0x02<<(8+CodeShift))
-#define PrefixREX            (0x04<<(8+CodeShift))
-#define Prefix38             (0x08<<(8+CodeShift))
-#define Prefix3A             (0x10<<(8+CodeShift))
-#define HasExtraFlags        (0x20<<(8+CodeShift))
-#define HasModRM             (0x40<<(8+CodeShift))
-#define ModRMShift           (7+8+CodeShift)
-#define SIBScaleShift        (ModRMShift+8-6)
-#define RegDWordDisplacement (0x01<<(8+SIBScaleShift))
-#define AddressMode          (0x02<<(8+SIBScaleShift))
-#define TypeShift            (2+8+SIBScaleShift)
-#define CategoryShift        5
-#define CategoryMask         ((1<<CategoryShift)-1)
-#define ModRM_mod            0xC0
-#define ModRM_reg            0x38
-#define ModRM_rm             0x07
-#define SIB_scale            0xC0
-#define SIB_index            0x38
-#define SIB_base             0x07
-#define REX_w                0x08
-
-#define MinRequired          8 // minimum required consecutive valid instructions to be considered as code
-
-
-class exeModel1: public Model {
-  BlockData& x;
-  Buf& buf;
-  const int N1, N2;
- ContextMap2 cm;
- IndirectMap iMap;
-    OpCache Cache;
-    ExeState pState , State ;
-    Instruction Op;
-    U32 TotalOps, OpMask, OpCategMask, Context, BrkPoint , BrkCtx ;
-    bool Valid;
-    U32 StateBH[256];
-    inline bool IsInvalidX64Op(U8 Op){
-  for (int i=0; i<19; i++){
-    if (Op == InvalidX64Ops[i])
-      return true;
-  }
-  return false;
-}
-
-inline bool IsValidX64Prefix(U8 Prefix){
-  for (int i=0; i<8; i++){
-    if (Prefix == X64Prefixes[i])
-      return true;
-  }
-  return ((Prefix>=0x40 && Prefix<=0x4F) || (Prefix>=0x64 && Prefix<=0x67));
-}
-
-void ProcessMode(Instruction &Op, ExeState &State){
-  if ((Op.Flags&fMODE)==fAM){
-    Op.Data|=AddressMode;
-    Op.BytesRead = 0;
-    switch (Op.Flags&fTYPE){
-      case fDR : Op.Data|=(2<<TypeShift);
-      case fDA : Op.Data|=(1<<TypeShift);
-      case fAD : {
-        State = Read32;
-        break;
-      }
-      case fBR : {
-        Op.Data|=(2<<TypeShift);
-        State = Read8;
-      }
-    }
-  }
-  else{
-    switch (Op.Flags&fTYPE){
-      case fBI : State = Read8; break;
-      case fWI : {
-        State = Read16;
-        Op.Data|=(1<<TypeShift);
-        Op.BytesRead = 0;
-        break;
-      }
-      case fDI : {
-        // x64 Move with 8byte immediate? [REX.W is set, opcodes 0xB8+r]
-        Op.imm8=((Op.REX & REX_w)>0 && (Op.Code&0xF8)==0xB8);
-        if (!Op.o16 || Op.imm8){
-          State = Read32;
-          Op.Data|=(2<<TypeShift);
-        }
-        else{
-          State = Read16;
-          Op.Data|=(3<<TypeShift);
-        }
-        Op.BytesRead = 0;
-        break;
-      }
-      default: State = Start; /*no immediate*/
-    }
-  }
-}
-
-void ProcessFlags2(Instruction &Op, ExeState &State){
-  //if arriving from state ExtraFlags, we've already read the ModRM byte
-  if ((Op.Flags&fMODE)==fMR && State!=ExtraFlags){
-    State = ReadModRM;
-    return;
-  }
-  ProcessMode(Op, State);
-}
-
-void ProcessFlags(Instruction &Op, ExeState &State){
-  if (Op.Code==OP_CALLF || Op.Code==OP_JMPF || Op.Code==OP_ENTER){
-    Op.BytesRead = 0;
-    State = Read16_f;
-    return; //must exit, ENTER has ModRM too
-  }
-  ProcessFlags2(Op, State);
-}
-
-void CheckFlags(Instruction &Op, ExeState &State){
-  //must peek at ModRM byte to read the REG part, so we can know the opcode
-  if (Op.Flags==fMEXTRA)
-    State = ExtraFlags;
-  else if (Op.Flags==fERR){
-    memset(&Op, 0, sizeof(Instruction));
-    State = Error;
-  }
-  else
-    ProcessFlags(Op, State);
-}
-
-void ReadFlags(Instruction &Op, ExeState &State){
-  Op.Flags = Table1[Op.Code];
-  Op.Category = TypeOp1[Op.Code];
-  CheckFlags(Op, State);
-}
-
-void ProcessModRM(Instruction &Op, ExeState &State){
-  if ((Op.ModRM & ModRM_mod)==0x40)
-    State = Read8_ModRM; //register+byte displacement
-  else if ((Op.ModRM & ModRM_mod)==0x80 || (Op.ModRM & (ModRM_mod|ModRM_rm))==0x05 || (Op.ModRM<0x40 && (Op.SIB & SIB_base)==0x05) ){
-    State = Read32_ModRM; //register+dword displacement
-    Op.BytesRead = 0;
-  }
-  else
-    ProcessMode(Op, State);
-}
-
-void ApplyCodeAndSetFlag(Instruction &Op, U32 Flag = 0){
-  Op.Data&=ClearCodeMask; \
-  Op.Data|=(Op.Code<<CodeShift)|Flag;
-}
-
-inline U32 OpN(OpCache &Cache, U32 n){
-  return Cache.Op[ (Cache.Index-n)&(CacheSize-1) ];
-}
-
-inline U32 OpNCateg(U32 &Mask, U32 n){
-  return ((Mask>>(CategoryShift*(n-1)))&CategoryMask);
-}
-
-inline int pref(int i) { return (buf(i)==0x0f)+2*(buf(i)==0x66)+3*(buf(i)==0x67); }
-
-// Get context at buf(i) relevant to parsing 32-bit x86 code
-U32 execxt(int i, int x=0) {
-  int prefix=0, opcode=0, modrm=0, sib=0;
-  if (i) prefix+=4*pref(i--);
-  if (i) prefix+=pref(i--);
-  if (i) opcode+=buf(i--);
-  if (i) modrm+=buf(i--)&(ModRM_mod|ModRM_rm);
-  if (i&&((modrm&ModRM_rm)==4)&&(modrm<ModRM_mod)) sib=buf(i)&SIB_scale;
-  return prefix|opcode<<4|modrm<<12|x<<20|sib<<(28-6);
-}
-public:
-  exeModel1(BlockData& bd,U32 val=0):x(bd),buf(bd.buf),N1(10), N2(10),cm(CMlimit(MEM()*4), N1+N2,M_EXE,
-  CM_MR+
-  CM_RUN0+
-  CM_MAIN1+
-  CM_MAIN2+
-  CM_MAIN3+
-  CM_MAIN4+
-  CM_M12+
-  CM_M6
-  ),iMap(20,1),
-  pState (Start), State( Start), TotalOps(0), OpMask(0),OpCategMask(0), Context(0),BrkPoint(0),
-  BrkCtx(0),Valid(false) {
-      memset(&Cache, 0, sizeof(OpCache));
-      memset(&Op, 0, sizeof(Instruction));
-      memset(&StateBH, 0, sizeof(StateBH));
-  }
-  int inputs() {return 2+(N1+N2)*cm.inputs();}
-  int nets() {return  1024+ 1024+ 1024+ 8192+ 8192+ 8192;}
-  int netcount() {return 6;}
-int p(Mixer& m,int val1=0,int val2=0){
-    if (x.filetype==DBASE ||x.filetype==HDR || x.filetype==ARM|| x.filetype==IMGUNK){
-        for (int i=0; i<inputs(); i++) m.add(0); 
-        return false;
-    }
-  if (x.bpos==0 ) {
-    pState = State;
-    U8 B = (U8)x.c4;
-    Op.Size++;
-    switch (State){
-      case Start: case Error: {
-        // previous code may have just been a REX prefix
-        bool Skip = false;
-        if (Op.MustCheckREX){
-          Op.MustCheckREX = false;
-          // valid x64 code?
-          if (!IsInvalidX64Op(B) && !IsValidX64Prefix(B)){
-            Op.REX = Op.Code;
-            Op.Code = B;
-            Op.Data = PrefixREX|(Op.Code<<CodeShift)|(Op.Data&PrefixMask); 
-            Skip = true;
-          }
-        }
-        
-        Op.ModRM = Op.SIB = Op.REX = Op.Flags = Op.BytesRead = 0;       
-        if (!Skip){
-          Op.Code = B;
-          // possible REX prefix?
-          Op.MustCheckREX = ((Op.Code&0xF0)==0x40) && (!(Op.Decoding && ((Op.Data&PrefixMask)==1)));
-          
-          // check prefixes
-          Op.Prefix = (Op.Code==ES_OVERRIDE || Op.Code==CS_OVERRIDE || Op.Code==SS_OVERRIDE || Op.Code==DS_OVERRIDE) + //invalid in x64
-                      (Op.Code==FS_OVERRIDE)*2 +
-                      (Op.Code==GS_OVERRIDE)*3 +
-                      (Op.Code==AD_OVERRIDE)*4 +
-                      (Op.Code==WAIT_FPU)*5 +
-                      (Op.Code==LOCK)*6 +
-                      (Op.Code==REP_N_STR || Op.Code==REP_STR)*7;
-
-          if (!Op.Decoding){
-            TotalOps+=(Op.Data!=0)-(Cache.Index && Cache.Op[ Cache.Index&(CacheSize-1) ]!=0);
-            OpMask = (OpMask<<1)|(State!=Error);
-            OpCategMask = (OpCategMask<<CategoryShift)|(Op.Category);
-            Op.Size = 0;
-            
-            Cache.Op[ Cache.Index&(CacheSize-1) ] = Op.Data;
-            Cache.Index++;
-
-            if (!Op.Prefix)
-              Op.Data = Op.Code<<CodeShift;
-            else{
-              Op.Data = Op.Prefix;
-              Op.Category = TypeOp1[Op.Code];
-              Op.Decoding = true;
-              BrkCtx = hash(1+(BrkPoint = 0), Op.Prefix, OpCategMask&CategoryMask);
-              break;
-            }
-          }
-          else{
-            // we only have enough bits for one prefix, so the
-            // instruction will be encoded with the last one
-            if (!Op.Prefix){
-              Op.Data|=(Op.Code<<CodeShift);
-              Op.Decoding = false;
-            }
-            else{
-              Op.Data = Op.Prefix;
-              Op.Category = TypeOp1[Op.Code];
-              BrkCtx = hash(1+(BrkPoint = 1), Op.Prefix, OpCategMask&CategoryMask);
-              break;
-            }
-          }
-        }
-
-        if ((Op.o16=(Op.Code==OP_OSIZE)))
-          State = Pref_Op_Size;
-        else if (Op.Code==OP_2BYTE)
-          State = Pref_MultiByte_Op;
-        else
-          ReadFlags(Op, State);
-        BrkCtx = hash(1+(BrkPoint = 2), State, Op.Code, (OpCategMask&CategoryMask), OpN(Cache,1)&((ModRM_mod|ModRM_reg|ModRM_rm)<<ModRMShift));
-        break;
-      }
-      case Pref_Op_Size : {
-        Op.Code = B;
-        ApplyCodeAndSetFlag(Op, OperandSizeOverride);
-        ReadFlags(Op, State);
-        BrkCtx = hash(1+(BrkPoint = 3), State);
-        break;
-      }
-      case Pref_MultiByte_Op : {
-        Op.Code = B;        
-        Op.Data|=MultiByteOpcode;
-
-        if (Op.Code==0x38)
-          State = Read_OP3_38;
-        else if (Op.Code==0x3A)
-          State = Read_OP3_3A;
-        else{
-          ApplyCodeAndSetFlag(Op);
-          Op.Flags = Table2[Op.Code];
-          Op.Category = TypeOp2[Op.Code];
-          CheckFlags(Op, State);
-        }
-        BrkCtx = hash(1+(BrkPoint = 4), State);
-        break;
-      }
-      case ParseFlags : {
-        ProcessFlags(Op, State);
-        BrkCtx = hash(1+(BrkPoint = 5), State);
-        break;
-      }
-      case ExtraFlags : case ReadModRM : {
-        Op.ModRM = B;
-        Op.Data|=(Op.ModRM<<ModRMShift)|HasModRM;
-        Op.SIB = 0;
-        if (Op.Flags==fMEXTRA){
-          Op.Data|=HasExtraFlags;
-          int i = ((Op.ModRM>>3)&0x07) | ((Op.Code&0x01)<<3) | ((Op.Code&0x08)<<1);
-          Op.Flags = TableX[i];
-          Op.Category = TypeOpX[i];
-          if (Op.Flags==fERR){
-            memset(&Op, 0, sizeof(Instruction));
-            State = Error;
-            BrkCtx = hash(1+(BrkPoint = 6), State);
-            break;
-          }
-          ProcessFlags(Op, State);
-          BrkCtx = hash(1+(BrkPoint = 7), State);
-          break;
-        }
-
-        if ((Op.ModRM & ModRM_rm)==4 && Op.ModRM<ModRM_mod){
-          State = ReadSIB;
-          BrkCtx = hash(1+(BrkPoint = 8), State);
-          break;
-        }
-
-        ProcessModRM(Op, State);
-        BrkCtx = hash(1+(BrkPoint = 9), State, Op.Code );
-        break;
-      }
-      case Read_OP3_38 : case Read_OP3_3A : {
-        Op.Code = B;
-        ApplyCodeAndSetFlag(Op, Prefix38<<(State-Read_OP3_38));
-        if (State==Read_OP3_38){
-          Op.Flags = Table3_38[Op.Code];
-          Op.Category = TypeOp3_38[Op.Code];
-        }
-        else{
-          Op.Flags = Table3_3A[Op.Code];
-          Op.Category = TypeOp3_3A[Op.Code];
-        }
-        CheckFlags(Op, State);
-        BrkCtx = hash(1+(BrkPoint = 10), State);
-        break;
-      }
-      case ReadSIB : {
-        Op.SIB = B;
-        Op.Data|=((Op.SIB & SIB_scale)<<SIBScaleShift);
-        ProcessModRM(Op, State);
-        BrkCtx = hash(1+(BrkPoint = 11), State, Op.SIB&SIB_scale);
-        break;
-      }
-      case Read8 : case Read16 : case Read32 : {
-        if (++Op.BytesRead>=((State-Read8)<<int(Op.imm8+1))){
-          Op.BytesRead = 0;
-          Op.imm8 = false;
-          State = Start;
-        }
-        BrkCtx = hash(1+(BrkPoint = 12), State, Op.Flags&fMODE, Op.BytesRead, ((Op.BytesRead>1)?(buf(Op.BytesRead)<<8):0)|((Op.BytesRead)?B:0) );
-        break;
-      }
-      case Read8_ModRM : {
-        ProcessMode(Op, State);
-        BrkCtx = hash(1+(BrkPoint = 13), State);
-        break;
-      }
-      case Read16_f : {
-        if (++Op.BytesRead==2){
-          Op.BytesRead = 0;
-          ProcessFlags2(Op, State);
-        }
-        BrkCtx = hash(1+(BrkPoint = 14), State);
-        break;
-      }
-      case Read32_ModRM : {
-        Op.Data|=RegDWordDisplacement;
-        if (++Op.BytesRead==4){
-          Op.BytesRead = 0;
-          ProcessMode(Op, State);
-        }
-        BrkCtx = hash(1+(BrkPoint = 15), State);
-        break;
-      }
-    }
-
-    Valid = (TotalOps>2*MinRequired) && ((OpMask&((1<<MinRequired)-1))==((1<<MinRequired)-1));
-    Context = State+16*Op.BytesRead+16*(Op.REX & REX_w);
-    StateBH[Context] = (StateBH[Context]<<8)|B;
-
-    if (Valid || val1){
-      int mask=0, count0=0;
-      int i=0;
-      while (i<N1){
-        if (i>1) mask=mask*2+(buf(i-1)==0), count0+=mask&1;
-        int j=(i<4)?i+1:5+(i-4)*(2+(i>6));
-        cm.set(hash(execxt(j, buf(1)*(j>6)), ((1<<N1)|mask)*(count0*N1/2>=i), (0x08|(x.blpos&0x07))*(i<4)));
-        i++;
-      }
-
-      cm.set(BrkCtx);
-      mask = PrefixMask|(0xF8<<CodeShift)|MultiByteOpcode|Prefix38|Prefix3A;
-      cm.set(hash(OpN(Cache, 1)&(mask|RegDWordDisplacement|AddressMode), State+16*Op.BytesRead, Op.Data&mask, Op.REX, Op.Category));
-      
-      mask = 0x04|(0xFE<<CodeShift)|MultiByteOpcode|Prefix38|Prefix3A|((ModRM_mod|ModRM_reg)<<ModRMShift);
-      cm.set(hash(
-        OpN(Cache, 1)&mask, OpN(Cache, 2)&mask, OpN(Cache, 3)&mask,
-        Context+256*((Op.ModRM & ModRM_mod)==ModRM_mod),
-        Op.Data&((mask|PrefixREX)^(ModRM_mod<<ModRMShift))
-      ));
-     
-      mask = 0x04|CodeMask;
-      cm.set(hash(OpN(Cache, 1)&mask, OpN(Cache, 2)&mask, OpN(Cache, 3)&mask, OpN(Cache, 4)&mask, (Op.Data&mask)|(State<<11)|(Op.BytesRead<<15)));
-
-      mask = 0x04|(0xFC<<CodeShift)|MultiByteOpcode|Prefix38|Prefix3A;
-      cm.set(hash(State+16*Op.BytesRead, Op.Data&mask, Op.Category*8 + (OpMask&0x07), Op.Flags, ((Op.SIB & SIB_base)==5)*4+((Op.ModRM & ModRM_reg)==ModRM_reg)*2+((Op.ModRM & ModRM_mod)==0)));
-
-      mask = PrefixMask|CodeMask|OperandSizeOverride|MultiByteOpcode|PrefixREX|Prefix38|Prefix3A|HasExtraFlags|HasModRM|((ModRM_mod|ModRM_rm)<<ModRMShift);
-      cm.set(hash(Op.Data&mask, State+16*Op.BytesRead, Op.Flags));
-
-      mask = PrefixMask|CodeMask|OperandSizeOverride|MultiByteOpcode|Prefix38|Prefix3A|HasExtraFlags|HasModRM;
-      cm.set(hash(OpN(Cache, 1)&mask, State, Op.BytesRead*2+((Op.REX&REX_w)>0), Op.Data&((U16)(mask^OperandSizeOverride))));
-
-      mask = 0x04|(0xFE<<CodeShift)|MultiByteOpcode|Prefix38|Prefix3A|(ModRM_reg<<ModRMShift);
-      cm.set(hash(OpN(Cache, 1)&mask, OpN(Cache, 2)&mask, State+16*Op.BytesRead, Op.Data&(mask|PrefixMask|CodeMask)));
-
-      cm.set(hash(++i,State+16*Op.BytesRead));
-
-      cm.set(hash(
-        (0x100|B)*(Op.BytesRead>0),
-        State+16*pState+256*Op.BytesRead,
-        ((Op.Flags&fMODE)==fAM)*16 + (Op.REX & REX_w) + (Op.o16)*4 + ((Op.Code & 0xFE)==0xE8)*2 + ((Op.Data & MultiByteOpcode)!=0 && (Op.Code & 0xF0)==0x80)
-      ));
-      
-    }
-  }
-  if (Valid || val1){
-    int  cnx=m.nx;
-    cm.mix(m, 1,4);
-    int count=(m.nx-cnx)/(N1+N2);
-    m.nx=cnx;
-    for (int i=count*(N1+N2);i!=0;--i) m.sp(5); // increase all
-
-    iMap.set(hash(BrkCtx, x.bpos));
-    if(x.filetype==EXE ) iMap.mix(m,1,4); 
-    else m.add(0),m.add(0);
-    x.x86_64.state = 0x80u | (static_cast<std::uint8_t>(State) << 3u) | Op.BytesRead;
-  }else{
-      x.x86_64.state=0;
-      for (int i=0; i<inputs(); ++i)
-        m.add(0);
-  }
-  U8 s = ((StateBH[Context]>>(28-x.bpos))&0x08) |
-         ((StateBH[Context]>>(21-x.bpos))&0x04) |
-         ((StateBH[Context]>>(14-x.bpos))&0x02) |
-         ((StateBH[Context]>>( 7-x.bpos))&0x01) |
-         ((Op.Category==OP_GEN_BRANCH)<<4)|
-         (((x.c0&((1<<x.bpos)-1))==0)<<5);
-
-  m.set((Context*4+(s>>4)), 1024);
-  m.set((State*64+x.bpos*8+(Op.BytesRead>0)*4+(s>>4)), 1024);
-  m.set( (BrkCtx&0x1FF)|((s&0x20)<<4), 1024 );
-  m.set(hash(Op.Code, State, OpN(Cache, 1)&CodeMask)&0x1FFF, 8192);
-  m.set(hash(State, x.bpos, Op.Code, Op.BytesRead)&0x1FFF, 8192);
-  m.set(hash(State, (x.bpos<<2)|(x.c0&3), OpCategMask&CategoryMask, ((Op.Category==OP_GEN_BRANCH)<<2)|(((Op.Flags&fMODE)==fAM)<<1)|(Op.BytesRead>0))&0x1FFF, 8192);
-  
-  return Valid;
-}
-
-
-virtual ~exeModel1(){ }
-};
 
 
 //////////////////////////// indirectModel /////////////////////
